requirements.txt
langchain
langchain-core
langchain-community
langchain-anthropic
langchain-huggingface
langchain-chroma
langchain-text-splitters
pypdf
python-dotenv
sentence-transformers
gradio
anthropic